requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.0,<3.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
stripe>=7.0,<8.0
PyJWT>=2.8,<3.0
httpx>=0.27,<1.0
